Added the ability to only view task detail view
If user copy the url of a task master detail view and send it to another user, it wouldn't  have work. It's because the url is part of the data subscription id. Now this new commit will redirect to a task detail only page if the subscription and the chosen task id are not compatible.